chore: move modules to core folder
Moved get_password, progress_bar, size_calculator from utils to core for better code structure maintability
