+++ src/autotarcompress/backup_manager.py
@@ -15,8 +15,8 @@ from typing import TYPE_CHECKING
 
 from autotarcompress.metadata import update_backup_metadata
 from autotarcompress.utils.hash_utils import calculate_sha256
-from autotarcompress.utils.progress_bar import SimpleProgressBar
-from autotarcompress.utils.size_calculator import SizeCalculator
+from autotarcompress.progress_bar import SimpleProgressBar
+from autotarcompress.size_calculator import SizeCalculator
 from autotarcompress.utils.utils import (
     ensure_backup_folder,
     validate_and_expand_paths,
