+++ src/autotarcompress/progress_bar.py
