+++ src/autotarcompress/utils/__init__.py
@@ -1,7 +1,5 @@
 """Utility package for backup operations."""
 
-from .get_password import PasswordContext
-from .size_calculator import SizeCalculator
 from .utils import (
     ensure_backup_folder,
     is_pv_available,
@@ -9,8 +7,6 @@ from .utils import (
 )
 
 __all__ = [
-    "PasswordContext",
-    "SizeCalculator",
     "ensure_backup_folder",
     "is_pv_available",
     "validate_and_expand_paths",
